move serialization related file into cocos/serialization
