+++ cocos/animation/marionette/state.ts
@@ -5,7 +5,7 @@ import { EditorExtendable } from '../../core/data/editor-extendable';
 import { CLASS_NAME_PREFIX_ANIM } from '../define';
 import { StateMachineComponent } from './state-machine-component';
 import { js } from '../../core';
-import { instantiate } from '../../core/data/instantiate';
+import { instantiate } from '../../serialization/instantiate';
 import { editorExtrasTag } from '../../core/data';
 import { cloneAnimationGraphEditorExtrasFrom } from './animation-graph-editor-extras-clone-helper';
