+++ cocos/asset/asset-manager/deserialize.ts
@@ -26,7 +26,7 @@
 import { EDITOR } from 'internal:constants';
 import { Asset } from '../assets/asset';
 import { MissingScript } from '../../misc/missing-script';
-import { deserialize, Details } from '../../core/data/deserialize';
+import { deserialize, Details } from '../../serialization/deserialize';
 import { error } from '../../core/platform/debug';
 import { js } from '../../core/utils/js';
 import { dependMap, nativeDependMap } from './depend-maps';
