+++ cocos/asset/asset-manager/pack-manager.ts
@@ -25,7 +25,7 @@
 
 import { ImageAsset } from '../assets/image-asset';
 import { Texture2D } from '../assets/texture-2d';
-import { packCustomObjData, unpackJSONs } from '../../core/data/deserialize';
+import { packCustomObjData, unpackJSONs } from '../../serialization/deserialize';
 import { error, errorID } from '../../core/platform/debug';
 import { js } from '../../core/utils/js';
 import Cache from './cache';
