+++ cocos/asset/asset-manager/parser.ts
@@ -32,7 +32,7 @@ import plistParser from './plist-parser';
 import { CompleteCallback, IDownloadParseOptions, files, parsed } from './shared';
 
 import { PixelFormat } from '../assets/asset-enum';
-import { CCON } from '../../core/data/ccon';
+import { CCON } from '../../serialization/ccon';
 import { Asset } from '../assets';
 
 // PVR constants //
