+++ cocos/core/data/class.ts
@@ -40,7 +40,15 @@ import { PropertyStash, PropertyStashInternalFlag } from './class-stash';
 
 const DELIMETER = attributeUtils.DELIMETER;
 const CCCLASS_TAG = '__ctors__'; // Still use this historical name to avoid unsynchronized version issue
+
+/**
+ * @engineInternal
+ */
 export const ENUM_TAG = 'Enum';
+
+/**
+ * @engineInternal
+ */
 export const BITMASK_TAG = 'BitMask';
 
 function pushUnique (array, item) {
@@ -354,6 +362,7 @@ CCClass.attr = attributeUtils.attr;
  * Returns if the class is a cc-class or is fast defined.
  * @param constructor The constructor of the class.
  * @returns Judge result.
+ * @engineInternal
  */
 // eslint-disable-next-line @typescript-eslint/ban-types
 export function isCCClassOrFastDefined<T> (constructor: Constructor<T>) {
