--- cocos/core/data/deserialize-dynamic-empty.ts
@@ -1,8 +1,0 @@
-
-export function deserializeDynamic() {
-    throw new Error(`Should not called`);
-}
-
-export function parseUuidDependenciesDynamic() {
-    throw new Error(`Should not called`);
-}
