--- cocos/core/data/deserialize-symbols.ts
@@ -1,1 +1,0 @@
-export const onAfterDeserializedTag = Symbol('[[OnAfterDeserialized]]');
