+++ cocos/core/data/editor-extendable.ts
@@ -1,7 +1,6 @@
 import { EDITOR } from 'internal:constants';
 import { ccclass, editorOnly } from 'cc.decorator';
 import { js } from '../utils/js';
-import { CCClass } from './class';
 import { EditorExtendableObject, editorExtrasTag } from './editor-extras-tag';
 import { assertIsTrue } from './utils/asserts';
