+++ cocos/core/data/index.ts
@@ -28,20 +28,24 @@ import { legacyCC } from '../global-exports';
 
 legacyCC._decorator = _decorator;
 export { _decorator };
-export { CCClass } from './class';
-export { CCObject, isValid } from './object';
-export { deserialize } from './deserialize';
-export { Details } from './deserialize';
-export { getSerializationMetadata } from './serialization-metadata';
-export type { SerializationMetadata } from './serialization-metadata';
-export { instantiate } from './instantiate';
+export { CCClass, ENUM_TAG, BITMASK_TAG, isCCClassOrFastDefined } from './class';
+export { CCObject, isCCObject, isValid } from './object';
 export { CCInteger, CCFloat, CCBoolean, CCString } from './utils/attribute';
+/**
+ * @engineInternal
+ */
+export * as Attr from './utils/attribute';
 export { CompactValueTypeArray } from './utils/compact-value-type-array';
 export { editorExtrasTag } from './editor-extras-tag';
+export * from './utils/asserts';
+export * from './utils/compiler';
 export { deserializeTag, serializeTag } from './custom-serializable';
 export type {
     SerializationInput,
     SerializationOutput,
     SerializationContext,
     CustomSerializable,
+    DeserializationContext,
 } from './custom-serializable';
+export { getSerializationMetadata } from './serialization-metadata';
+export type { SerializationMetadata } from './serialization-metadata';
