+++ cocos/core/data/utils/asserts.ts
@@ -29,6 +29,7 @@ import { DEBUG } from 'internal:constants';
  * Asserts that the expression is non-nullable, i.e. is neither `null` nor `undefined`.
  * @param expr Testing expression.
  * @param message Optional message.
+ * @engineInternal
  */
 export function assertIsNonNullable<T> (expr: T, message?: string): asserts expr is NonNullable<T> {
     assertIsTrue(!(expr === null || expr === undefined), message);
@@ -38,6 +39,7 @@ export function assertIsNonNullable<T> (expr: T, message?: string): asserts expr
  * Asserts that the expression evaluated to `true`.
  * @param expr Testing expression.
  * @param message Optional message.
+ * @engineInternal
  */
 export function assertIsTrue (expr: unknown, message?: string): asserts expr {
     if (DEBUG && !expr) {
@@ -47,6 +49,10 @@ export function assertIsTrue (expr: unknown, message?: string): asserts expr {
     }
 }
 
+/**
+ * Assets that the index is valid.
+ * @engineInternal
+ */
 export function assertsArrayIndex<T> (array: T[], index: number) {
     assertIsTrue(index >= 0 && index < array.length, `Array index ${index} out of bounds: [0, ${array.length})`);
 }
