+++ cocos/core/data/utils/attribute.ts
@@ -40,6 +40,7 @@ export function createAttrsSingle (owner: Object, superAttrs?: any) {
 
 /**
  * @param subclass Should not have '__attrs__'.
+ * @engineInternal
  */
 export function createAttrs (subclass: any) {
     if (typeof subclass !== 'function') {
@@ -86,6 +87,7 @@ export function attr (constructor: any, propertyName: string): { [attributeName:
 
 /**
  * Returns a readonly meta object.
+ * @engineInternal
  */
 export function getClassAttrs (constructor: any) {
     return (constructor.hasOwnProperty('__attrs__') && constructor.__attrs__) || createAttrs(constructor);
