+++ tests/core/editor-extendable.test.ts
@@ -1,4 +1,5 @@
-import { EditorExtendable, editorExtrasTag } from "../../cocos/core/data/editor-extendable";
+import { EditorExtendable } from "../../cocos/core/data/editor-extendable";
+import { editorExtrasTag } from "../../cocos/core/data/editor-extras-tag";
 
 describe('Editor extendable', () => {
     test('Serialize', () => {
