+++ tests/core/serialization/deserialize-missing-class.test.ts
@@ -1,6 +1,6 @@
 
 import { MissingScript } from '../../../cocos/misc';
-import { deserialize } from '../../../cocos/core/data/deserialize';
+import { deserialize } from '../../../cocos/serialization/deserialize';
 import { js } from '../../../cocos/core/utils/js';
 
 describe(`Missing class deserialization`, () => {
